Preserve withheld values in output data per #935
The one slightly weird thing here is that we're converting withheld
values (indicated by "W") in both the original ONRR data and our
intermediate (county, state, and offshore) files into _exactly_ 0.00 in
the TSVs that the front end consumes. This means that we can still sum
them up and not get NaN issues, but we can also detect exactly zero and
show "(withheld)" if that's what we get for a given geographical rollup.