Updates open/close behaviors for glossary terms
Closes #1563

Previously, when opening the gallery and reading through the list of
terms, one had to click on the chevron next to a term itself in order
to open or close the definition.

This patch allows for performing the same action but also through
clicking on the term itself, expanding the click area and making for
what I feel is a more expected behavior.

**Caveats**
 - I'm not familiar with the codebase, so while this seems to _work_, I
   don't know if it's _right_ for the project, also, I wasn't even sure
   if I needed to submit the changes to the minified code in this PR.

 - The actual work seems like it's happening inside of the `List()`
   which made it more difficult to figure out. This is also why I'm not
   convinced that I made the right changes in the right places.

 - There are lingering issues with the cursor: should it be a pointer
   when hovering over the glossary terms?

Thanks!