Pin package dependencies to specific versions
In #1565 I ran into an issue where changes were brought in to the
minified javascript bundle that were unrelated to the changes made in
the PR. This was due to NPM packages which were pulled in under
the compatability version spec `pkg: ^1.2.3` but which differed from the
versions used to build the current release in `dev`.

This PR only pins the versions of the three libraries that interfered
with #1563, but it might be worthwhile to ammend it and pin all of the
dependency versions to ensure that all development environments are
building with the same code.

Cheers!