Update redirect behavior to include explanatory message.
Previously the user would receive something akin to a FOUC (flash of
unstyled content) except that it actually was styled, but for the 404
page.

Figuring that this wasn't all to helpful, I updated the behavior of the
redirect to include a message indication that the request page was not
found and also that the browser would automatically redirect after a
delay. A default delay of five seconds was chosen.

Now, when visiting an unknown page, an error message displays under the
large 404 message indicating that the request path does not exist,
provides a link to the redirecting path, and displays a countdown until
the redirect occurs.