clean up state names in coal all production
The source data (from EIA) for all lands coal production contains state
names like "Kentucy (Northern)" and "Pennsylvania (Anthracite)". This
change updates the transform for coal data to strip the parentheticals,
and commits the newly rendered data. I discovered this issue while
investigating #1644, but this is not a full fix.