Adjust mobile display styling for "diving previews"
Resolves #988

Previously it was noted that the graphics for the different resource
types on the [How It Works page](http://localhost:4000/how-it-works/)
look funny on narrow screens. Because of the width of the displays, the
text column was too narrow.

This patch incorporates @ericonne's suggestion for an adjustment on
screens smaller than a certain width (in this case, `tiny-down`).
Included here are cropped icons from the corresponding "diving"
background images.

**Caveats**
 - The CSS changes I introduced appear to get-the-job-done but someone
   with real skillz should probaby look over the changes and suggest a
   better way to accomplish it. I'm not confident that this is good
   styling the way I have written it.

 - Because the mockup showed the landing header beneath the icons but
   the header HTML is outside of the scope that the icon lives in, I
   have used _creative margins_ to swap their spots vertically instead
   of doing something like rendering the icon as an `<img />` lexically
   above the heading while dropping the existing background image
   reference entirely (for small screens).

 - The text in all of the mockups appears to be of a lighter-gray color
   while the current text in `dev` is black. I have not addressed this
   since it appeared ancillary to the actual issue.