Use Mercator projection for states and offshore regions
Previously, every projection for the data explorer was set to a custom
Albers conical projection.

This change swaps out to Mercator for the state and offshort region
maps. The Mercator should not only horizontally level the smaller
regions, but also produce something more familiar at the larger scales
than the Albers.