Add minimum Ruby version requirement
Previously, while attempting to build the site, I could not get the
build to complete without error. It complained of `to_h` not existing.

After tracking down this problem, I realized that my system was running
a late Ruby 2.0-release and `to_h` requires `2.1`

This patch adds a specifier in the Gemfile to require a minimum version.