Factor `year` out of commodity table component
Previously, more complete data was being sent into the components which
generate the individual commodity tables. This was needless, however,
and only added more coupling to the React components.

The year has been refactored out of the commodity components and moved
into the parent table component. Hopefully this is just the first step
in understanding the data better and continuing to decouple from the
components.