Break apart subtypes from types
Previously, when focussing just on oil and gas for development purposes,
each row in the output table was its own `<CommodityTable />` but this
doesn't match the structure that the page expects.

Now, all of the subtypes of commodities are handed down to the commodity
table so that there can be one header for all of the subtypes.