Show Oil and Gas table as distinct from others
Previously, the gas and oil tables were being generated as all of the
other commodity type tables. Unfortunately, in the existing Jekyll
rendering, the gas and oil tables form from a custom aggregate of oil
and gas data.

This patch creates a new `<OilGasTable />` component to handle the
rendering. It also adds the functionality in `<FilledBar />` to render
multiple values.