Sticky h3 DEMO (not for merging)
Used @shawnbot’s brilliant sticky code to create a rough demo showing
how it might be applied on the explore pages. The superfluous padding
above the sticky heads will need to be eliminated, but that’s beyond my
ken.

As a hack attempt, this is definitely for Demonstration Purposes Only.
I hope it’s useful.