Various table format adjustments
 - Equalized top and bottom cell padding
 - Changed indenting method to padding-left
 - Added styling for the “sup” class, which enabled me to shift the
superscript numbers slightly down, as well as adding a little spacing
on the left (so they don’t almost hit the adjacent text)
 - Tightened up the table line height a bit

On these tables, we still need to adjust the leading for lines which
have footnotes. (The number forces line heights to be too generous.)

Definitely high likelihood of syntax glitches here! :fingers crossed:

More to come!