Setup basic fractal styleguide
The styleguide is in the `nrrd-design-system` folder. You can watch
changes to the styleguide with `watch-styleguide` which gives you
a local server to use. On federalist, the styleguide is built into
static files and moved from `nrrd-design-system/public` to
`./styleguide` where it's served by jekyll. So the fractal site will
be available at `...gov/styleguide`.