Use webpack to compile Sass
Because of the way Federalist is set up, we don't have access to Jekyll
at the same time that the federalist node task is run. We want to have
access to that for Fractal so that we can make Federalist preview links
for the style guide branches.

Therefore, this moves the sass compilation to webpack and out of
Jekyll-land. Now that the CSS asset compilation is handled by JS, it
will definitely be available when it comes time to run the 'federalist'
build task.

It partially does this by outputting files to /public instead of /css
and /js/lib and so many of these file changes are to reflect this new
path