Use different _config.yml for dev and federalist
During development, we use a system of docker,
express, and jekyll to make the environment
run on different machines. However, in doing so
we told jekyll to ignore the /public directory
so that it could be managed by webpack &
express. On Federalist, there is no Express
server so jekyll needs to include /public when
it builds the site.