Use a base docker image with ruby and node
We need both ruby and node but using a base ruby image and installing
node everytime was a bit cumbersome. CircleCI publishes some handy
docker images and one has ruby and node.

In addition, docker installs the dependencies in the project image so
that the user shouldn't have to know to do that (refs #2619)