Fix homepage maps that broke in Jekyll upgrade
The homepage maps broke when we did a Jekyll upgrade because of some
change in Jekyll behavior. When we make Jekyll collections, the `id`
attribute gets treated differently.

From the docs (https://jekyllrb.com/docs/variables/):
page.id - An identifier unique to a document in a Collection or a Post
(useful in RSS feeds). e.g.  /2008/12/14/my-post
/my-collection/my-document

This matches the behavior since the generated HTML had a lot of URLs in
it instead of state abbreviations so I tried `state.slug` and
`region.slug` and it seems like it worked.