Add utility context variables
Some components need a special preview to display properly. In this case, the
callout cards need a non-white background so it is easier to see them. Instead
of including the container in the component template, add utility context
variables to control the preview layout and add the appropriate container
element.