Include only documented components
Re-write the pattern library query to iterate over README.md documents in
src/components. That way only components that have a README will be included in
the pattern library.

We use `onCreateNode` to create a componentId that we can use to easily link
README document nodes to docgen nodes that come from JS files.