Copy lato and eiti font assets
When importing the main.css into Gatsby, we run into issues with the paths being
incorrect in lato.css and eiti.css for their font assets. These assets were not
being copied to the build, so the rules were effectively broken. These are
fallbacks anyway, so they probably went unnoticed in production.